Fix broadcasting rules given upstream broadcasting fixes
There was a bug in Julia broadcasting rules fixed by https://github.com/JuliaLang/julia/pull/35948, but that fix breaks a test here, since this package was inadvertantly relying on the old incorrect logic. This fixes this package, basically correctly implements the API, and doesn't break it on current Julia versions either.